update predict code
