+++ examples/yinglong/predict_12layers.py
@@ -17,83 +17,11 @@ from os import path as osp
 
 import h5py
 import numpy as np
-import paddle
-import paddle.inference as paddle_infer
 import pandas as pd
-from packaging import version
 
-from examples.yinglong.timefeatures import time_features
+from examples.yinglong.plot import save_plot_weather_from_dict
+from examples.yinglong.predictor import YingLong
 from ppsci.utils import logger
-from ppsci.visualize import save_plot_weather_from_dict
-
-
-class YingLong:
-    def __init__(
-        self, model_file: str, params_file: str, mean_path: str, std_path: str
-    ):
-        self.model_file = model_file
-        self.params_file = params_file
-
-        config = paddle_infer.Config(model_file, params_file)
-        config.switch_ir_optim(False)
-        config.enable_use_gpu(100, 0)
-        config.enable_memory_optim()
-
-        self.predictor = paddle_infer.create_predictor(config)
-
-        # get input names and data handles
-        self.input_names = self.predictor.get_input_names()
-        self.input_data_handle = self.predictor.get_input_handle(self.input_names[0])
-        self.time_stamps_handle = self.predictor.get_input_handle(self.input_names[1])
-        self.nwp_data_handle = self.predictor.get_input_handle(self.input_names[2])
-
-        # get output names and data handles
-        self.output_names = self.predictor.get_output_names()
-        self.output_handle = self.predictor.get_output_handle(self.output_names[0])
-
-        # load mean and std data
-        self.mean = np.load(mean_path).reshape(-1, 1, 1).astype(np.float32)
-        self.std = np.load(std_path).reshape(-1, 1, 1).astype(np.float32)
-
-    def _preprocess_data(self, input_data, time_stamps, nwp_data):
-        # normalize data
-        input_data = (input_data - self.mean) / self.std
-        nwp_data = (nwp_data - self.mean) / self.std
-
-        # process time stamps
-        for i in range(len(time_stamps)):
-            time_stamps[i] = pd.DataFrame({"date": time_stamps[i]})
-            time_stamps[i] = time_features(time_stamps[i], timeenc=1, freq="h").astype(
-                np.float32
-            )
-        time_stamps = np.asarray(time_stamps)
-        return input_data, time_stamps, nwp_data
-
-    def _postprocess_data(self, data):
-        # denormalize data
-        data = data * self.std + self.mean
-        return data
-
-    def __call__(self, input_data, time_stamp, nwp_data):
-        # preprocess data
-        input_data, time_stamps, nwp_data = self._preprocess_data(
-            input_data, time_stamp, nwp_data
-        )
-
-        # set input data
-        self.input_data_handle.copy_from_cpu(input_data)
-        self.time_stamps_handle.copy_from_cpu(time_stamps)
-        self.nwp_data_handle.copy_from_cpu(nwp_data)
-
-        # run predictor
-        self.predictor.run()
-
-        # get predict data
-        pred_data = self.output_handle.copy_to_cpu()
-
-        # postprocess data
-        pred_data = self._postprocess_data(pred_data)
-        return pred_data
 
 
 def parse_args():
@@ -101,31 +29,31 @@ def parse_args():
     parser.add_argument(
         "--model_file",
         type=str,
-        default="./yinglong_models/yinglong_12.pdmodel",
+        default="../yinglong_models/yinglong_12.pdmodel",
         help="Model filename",
     )
     parser.add_argument(
         "--params_file",
         type=str,
-        default="./yinglong_models/yinglong_12.pdiparams",
+        default="../yinglong_models/yinglong_12.pdiparams",
         help="Parameter filename",
     )
     parser.add_argument(
         "--mean_path",
         type=str,
-        default="./hrrr_example_24vars/stat/mean_crop.npy",
+        default="../hrrr_example_24vars/stat/mean_crop.npy",
         help="Mean filename",
     )
     parser.add_argument(
         "--std_path",
         type=str,
-        default="./hrrr_example_24vars/stat/std_crop.npy",
+        default="../hrrr_example_24vars/stat/std_crop.npy",
         help="Standard deviation filename",
     )
     parser.add_argument(
         "--input_file",
         type=str,
-        default="./hrrr_example_24vars/valid/2022/01/01.h5",
+        default="../hrrr_example_24vars/valid/2022/01/01.h5",
         help="Input filename",
     )
     parser.add_argument(
@@ -134,7 +62,7 @@ def parse_args():
     parser.add_argument(
         "--nwp_file",
         type=str,
-        default="./hrrr_example_24vars/nwp_convert/2022/01/01/00.h5",
+        default="../hrrr_example_24vars/nwp_convert/2022/01/01/00.h5",
         help="NWP filename",
     )
     parser.add_argument(
@@ -150,13 +78,6 @@ def parse_args():
 def main():
     args = parse_args()
     logger.init_logger("ppsci", osp.join(args.output_path, "predict.log"), "info")
-    if version.Version(paddle.__version__) != version.Version("2.5.2"):
-        logger.error(
-            f"Your Paddle version is {paddle.__version__}, but this code currently "
-            "only supports PaddlePaddle 2.5.2. The latest version of Paddle will be "
-            "supported as soon as possible."
-        )
-        exit()
 
     num_timestamps = args.num_timestamps
 
@@ -173,11 +94,11 @@ def main():
     input_file = h5py.File(args.input_file, "r")["fields"]
     nwp_file = h5py.File(args.nwp_file, "r")["fields"]
 
-    # input_data.shape: (1, num_vars, 440, 408), num_vars = 24 for 12 layers
+    # input_data.shape: (1, 24, 440, 408)
     input_data = input_file[0:1]
-    # nwp_data.shape: # (num_timestamps, num_vars, 440, 408)
+    # nwp_data.shape: # (num_timestamps, 24, 440, 408)
     nwp_data = nwp_file[0:num_timestamps]
-    # ground_truth.shape: (num_timestamps, num_vars, 440, 408)
+    # ground_truth.shape: (num_timestamps, 24, 440, 408)
     ground_truth = input_file[1 : num_timestamps + 1]
 
     # create time stamps
@@ -189,7 +110,7 @@ def main():
 
     # run predictor
     pred_data = predictor(input_data, time_stamps, nwp_data)
-    pred_data = pred_data.squeeze(axis=1)  # (num_timestamps, num_vars, 440, 408)
+    pred_data = pred_data.squeeze(axis=1)  # (num_timestamps, 24, 440, 408)
 
     # save predict data
     save_path = osp.join(args.output_path, "result.npy")
@@ -205,12 +126,12 @@ def main():
     data_dict = {}
     visu_keys = []
     for i in range(num_timestamps):
-        visu_key = f"Init time: {args.init_time}h, YingLong pred: {i+1}h"
-        visu_keys.append(visu_key)
-        data_dict[visu_key] = pred_wind[i]
-        visu_key = f"Init time: {args.init_time}h, Ground truth: {i+1}h"
+        visu_key = f"Init time: {args.init_time}h\n Ground truth: {i+1}h"
         visu_keys.append(visu_key)
         data_dict[visu_key] = ground_truth_wind[i]
+        visu_key = f"Init time: {args.init_time}h\n YingLong-12 Layers: {i+1}h"
+        visu_keys.append(visu_key)
+        data_dict[visu_key] = pred_wind[i]
 
     save_plot_weather_from_dict(
         foldername=args.output_path,
